+++ lib/ndr_dev_support/integration_testing/drivers/chrome_headless.rb
@@ -1,4 +1,3 @@
-require 'chromedriver-helper'
 require 'selenium-webdriver'
 
 Capybara.register_driver(:chrome_headless) do |app|